# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:3001
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /user/register:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    post:
      tags:
        - User
      description: Register a new User
      # used as the method name of the controller
      operationId: registerUser
      parameters:
        - in: body
          name: register data
          description: The data to register a new user
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/TokenResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/autheticate:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    post:
      tags:
        - User
      description: Authenticate a new User
      # used as the method name of the controller
      operationId: authenticateUser
      parameters:
        - in: body
          name: request
          description: The data to authenticate a new user
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/TokenResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
tags:
  - name: User
    description: Resource for User data
definitions:
  TokenResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
      err:
        type: object